+++ NEWS.md
@@ -1,14 +1,28 @@
-# FlickrAPI 0.1.0.2 (2020-10-12)
-New feature updates to getPhotoSearch() by [@elipousson](https://github.com/elipousson).
+## FlickrAPI 0.1.1 (2024-10-08)
+
+* Update `getPhotoSearch()` to allow searching by date taken and upload date.
+* Add secret parameter to `getExif()` function to allow sharing of individual photos without API key access.
+* Import [{cli}](https://github.com/r-lib/cli) package for error messages.
+* Add `error_call` parameter to `FlickrAPIRequest()`.
+* Add `strict` parameter to `getFlickrAPIKey()` to allow function to warn instead of error if API key can't be found.
+* Replace `match.arg()` with `rlang::arg_match()` internally to improve error messages.
+
+## FlickrAPI 0.1.0.2 (2020-10-12)
+
+* New feature updates to `getPhotoSearch()` by [@elipousson](https://github.com/elipousson).
 
 ## FlickrAPI 0.1.0.1 (2020-09-03)
-New feature added. getPhotoSearch() by [@fozy81](https://github.com/fozy81).
+
+* New feature added. `getPhotoSearch()` by [@fozy81](https://github.com/fozy81).
 
 ## FlickrAPI 0.1.0.01 (2020-05-04)
-Minor update. Fixed reported bugs, updated exsiting features.
+
+* Minor update. Fixed reported bugs, updated exsiting features.
 
 ## FlickrAPI 0.1.0.0 (2019-01-13)
-Initial release.
+
+* Initial release.
 
 ## FlickrAPI 0.0.0.9000 (2019-01-12)
-Beta version of FlickrAPI package. Available only on [GitHub](https://github.com/koki25ando/FlickrAPI).
+
+* Beta version of FlickrAPI package. Available only on [GitHub](https://github.com/koki25ando/FlickrAPI).